refactor: remove redundant int cast in SummaryRow aggregation
Quantity is already int so the explicit (int) cast on g.Sum(i => i.Quantity)
was unnecessary noise. Removing it keeps the LINQ projection clean.

+++ backend/src/Adapters/Anela.Heblo.Adapters.ShoptetApi/Expedition/ExpeditionProtocolDocument.cs
@@ -135,7 +135,7 @@ public class ExpeditionProtocolDocument : IDocument
                         Variant = first.Variant,
                         WarehousePosition = first.WarehousePosition,
                         Unit = first.Unit,
-                        TotalQuantity = (int)g.Sum(i => i.Quantity),
+                        TotalQuantity = g.Sum(i => i.Quantity),
                         StockCount = first.StockCount,
                         IsFromSet = first.IsFromSet,
                     };
